Move TypeSystem, ValueConverter, and TypeNameConvention to `:api`
