+++ api/src/main/kotlin/io/spine/protodata/type/TypeSystem.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.codegen
+package io.spine.protodata.type
 
 import io.spine.protodata.EnumType
 import io.spine.protodata.File
