+++ api/src/main/kotlin/io/spine/protodata/type/ValueConverter.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.codegen
+package io.spine.protodata.type
 
 import io.spine.protodata.Type
 import io.spine.protodata.Value
