+++ api/src/main/proto/spine/protodata/value.proto
